+++ lexer.c
@@ -57,8 +57,8 @@ void parse_paren(Lexer *lexer, Token *tok) {
 }
 
 void parse_string(Lexer *lexer, Token *tok) {
-	size_t start = lexer->ptr;
-	size_t len = 0;
+	u32 start = lexer->ptr;
+	u32 len = 0;
 	char prev = '\0'; // tracks the previous character to check escape seq
 
 	while (!end(lexer)) {
@@ -80,8 +80,8 @@ void parse_string(Lexer *lexer, Token *tok) {
 }
 
 void parse_number(Lexer *lexer, Token *tok) {
-	size_t start = lexer->ptr;
-	size_t len = 0;
+	u32 start = lexer->ptr;
+	u32 len = 0;
 	while (!end(lexer)) {
 		char c = peek(lexer);
 		if (!(isdigit(c) || c == '.'))
@@ -105,8 +105,8 @@ bool is_ident_char(char c) {
 }
 
 void parse_identifier(Lexer *lexer, Token *tok) {
-	size_t start = lexer->ptr;
-	size_t len = 0;
+	u32 start = lexer->ptr;
+	u32 len = 0;
 	while (!end(lexer)) {
 		char c = peek(lexer);
 		if (!is_ident_char(c))
@@ -120,13 +120,13 @@ void parse_identifier(Lexer *lexer, Token *tok) {
 	tok->type = TOK_IDENT;
 }
 
-bool lex(Token* tokens, size_t token_len, String *input) {
+bool lex(Token* tokens, u32 token_len, String *input) {
 	Lexer lexer = {
 		.buf = input,
 		.ptr = 0,
 	};
 
-	size_t i = 0; // iterator for available tokens
+	u32 i = 0; // iterator for available tokens
 	while (!end(&lexer)) {
 		// Assert that we have enough space in our Token buffer to parse.
 		assert(i < token_len);
