+++ parser.c
@@ -3,7 +3,7 @@
 #include "parser.h"
 #include "lexer.h"
 
-Expression *parse(Expression *expressions, size_t expressions_len, Token *tokens, size_t tokens_len) {
+Expression *parse(Expression *expressions, u32 expressions_len, Token *tokens, u32 tokens_len) {
 	Parser parser = (Parser) {
 		.expressions_buf = expressions,
 		.expressions_len = expressions_len,
@@ -24,7 +24,7 @@ Expression *parse(Expression *expressions, size_t expressions_len, Token *tokens
 	// NOTE: This is similar to the `parse_list` function since a program
 	// is also just a list of expressions.
 	while (parser_peek(&parser)->type != TOK_EOF) {
-		size_t len = program->data.list.length;
+		u32 len = program->data.list.length;
 		assert(len < program->data.list.capacity);
 
 		Token *token = parser_peek(&parser);
@@ -104,7 +104,7 @@ Expression *parse_list(Parser *parser) {
 	// TODO: Don't exceed grabbing tokens outside of bounds of token length.
 	while (true) {
 		// Ensure we have enough list capacity to parse this.
-		size_t len = expr->data.list.length;
+		u32 len = expr->data.list.length;
 		assert(expr->data.list.capacity > len);
 
 		Token *token = parser_peek(parser);
